revert unnecessary changes

+++ javaslang/src/main/java/javaslang/collection/Array.java
@@ -1327,7 +1327,7 @@ interface ArrayModule {
                 return Array.of(Array.empty());
             } else {
                 return elements.zipWithIndex().flatMap(
-                        t -> apply(elements.drop(t._2.intValue() + 1), (k - 1)).map(c -> c.prepend(t._1))
+                        t -> apply(elements.drop(t._2 + 1), (k - 1)).map(c -> c.prepend(t._1))
                 );
             }
         }
