+++ javaslang/src/main/java/javaslang/collection/List.java
@@ -1647,7 +1647,7 @@ interface ListModule {
                 return List.of(List.empty());
             } else {
                 return elements.zipWithIndex().flatMap(
-                        t -> apply(elements.drop(t._2.intValue() + 1), (k - 1)).map(c -> c.prepend(t._1))
+                        t -> apply(elements.drop(t._2 + 1), (k - 1)).map(c -> c.prepend(t._1))
                 );
             }
         }
