+++ javaslang/src/main/java/javaslang/collection/Stream.java
@@ -1627,7 +1627,7 @@ interface StreamModule {
                 return Stream.of(Stream.empty());
             } else {
                 return elements.zipWithIndex().flatMap(
-                        t -> apply(elements.drop(t._2.intValue() + 1), (k - 1)).map((Stream<T> c) -> c.prepend(t._1))
+                        t -> apply(elements.drop(t._2 + 1), (k - 1)).map((Stream<T> c) -> c.prepend(t._1))
                 );
             }
         }
