+++ javaslang/src/main/java/javaslang/collection/Vector.java
@@ -1332,7 +1332,7 @@ interface VectorModule {
             if (k == 0) {
                 return Vector.of(Vector.empty());
             } else {
-                return elements.zipWithIndex().flatMap(t -> apply(elements.drop(t._2.intValue() + 1), (k - 1))
+                return elements.zipWithIndex().flatMap(t -> apply(elements.drop(t._2 + 1), (k - 1))
                         .map((Vector<T> c) -> c.prepend(t._1))
                 );
             }
